requirements.txt
# Flask Backend
flask==3.0.0
flask-cors==4.0.0

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1
